+++ priam/src/test/java/com/netflix/priam/backup/TestCompression.java
@@ -17,8 +17,17 @@
 
 package com.netflix.priam.backup;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
 import com.netflix.priam.compress.SnappyCompression;
 import com.netflix.priam.utils.SystemUtils;
+import java.io.*;
+import java.util.Enumeration;
+import java.util.Iterator;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipFile;
+import java.util.zip.ZipOutputStream;
 import org.apache.commons.io.IOUtils;
 import org.junit.After;
 import org.junit.Before;
@@ -27,52 +36,42 @@ import org.junit.Test;
 import org.xerial.snappy.SnappyInputStream;
 import org.xerial.snappy.SnappyOutputStream;
 
-import java.io.*;
-import java.util.Enumeration;
-import java.util.Iterator;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipFile;
-import java.util.zip.ZipOutputStream;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-
 @Ignore("does this test really have to generate smoke to verify correct behavior?")
-public class TestCompression
-{
+public class TestCompression {
 
     @Before
-    public void setup() throws IOException
-    {
+    public void setup() throws IOException {
         File f = new File("/tmp/compress-test.txt");
-        try(FileOutputStream stream = new FileOutputStream(f)) {
+        try (FileOutputStream stream = new FileOutputStream(f)) {
             for (int i = 0; i < (1000 * 1000); i++) {
-                stream.write("This is a test... Random things happen... and you are responsible for it...\n".getBytes("UTF-8"));
-                stream.write("The quick brown fox jumps over the lazy dog.The quick brown fox jumps over the lazy dog.The quick brown fox jumps over the lazy dog.\n".getBytes("UTF-8"));
+                stream.write(
+                        "This is a test... Random things happen... and you are responsible for it...\n"
+                                .getBytes("UTF-8"));
+                stream.write(
+                        "The quick brown fox jumps over the lazy dog.The quick brown fox jumps over the lazy dog.The quick brown fox jumps over the lazy dog.\n"
+                                .getBytes("UTF-8"));
             }
         }
     }
 
     @After
-    public void done()
-    {
+    public void done() {
         File f = new File("/tmp/compress-test.txt");
-        if (f.exists())
-            f.delete();
+        if (f.exists()) f.delete();
     }
 
-    void validateCompression(String uncompress, String compress)
-    {
+    void validateCompression(String uncompress, String compress) {
         File uncompressed = new File(uncompress);
         File compressed = new File(compress);
         assertTrue(uncompressed.length() > compressed.length());
     }
 
     @Test
-    public void zip() throws IOException
-    {
+    public void zip() throws IOException {
         BufferedInputStream source = null;
-        try(ZipOutputStream out = new ZipOutputStream(new BufferedOutputStream(new FileOutputStream("/tmp/compressed.zip")))) {
+        try (ZipOutputStream out =
+                new ZipOutputStream(
+                        new BufferedOutputStream(new FileOutputStream("/tmp/compressed.zip")))) {
             byte data[] = new byte[2048];
             File file = new File("/tmp/compress-test.txt");
             FileInputStream fi = new FileInputStream(file);
@@ -91,12 +90,12 @@ public class TestCompression
     public void unzip() throws IOException {
         ZipFile zipfile = new ZipFile("/tmp/compressed.zip");
         Enumeration e = zipfile.entries();
-        while (e.hasMoreElements())
-        {
+        while (e.hasMoreElements()) {
             ZipEntry entry = (ZipEntry) e.nextElement();
-            try(BufferedInputStream is = new BufferedInputStream(zipfile.getInputStream(entry));
-                BufferedOutputStream dest1 = new BufferedOutputStream(new FileOutputStream("/tmp/compress-test-out-0.txt"), 2048)) {
-                ;
+            try (BufferedInputStream is = new BufferedInputStream(zipfile.getInputStream(entry));
+                    BufferedOutputStream dest1 =
+                            new BufferedOutputStream(
+                                    new FileOutputStream("/tmp/compress-test-out-0.txt"), 2048)) {;
                 int c;
                 byte d[] = new byte[2048];
 
@@ -111,15 +110,15 @@ public class TestCompression
     }
 
     @Test
-    public void snappyCompress() throws IOException
-    {
+    public void snappyCompress() throws IOException {
         FileInputStream fi = new FileInputStream("/tmp/compress-test.txt");
-        SnappyOutputStream out = new SnappyOutputStream(new BufferedOutputStream(new FileOutputStream("/tmp/test0.snp")));
+        SnappyOutputStream out =
+                new SnappyOutputStream(
+                        new BufferedOutputStream(new FileOutputStream("/tmp/test0.snp")));
         BufferedInputStream origin = new BufferedInputStream(fi, 1024);
         byte data[] = new byte[1024];
         int count;
-        while ((count = origin.read(data, 0, 1024)) != -1)
-        {
+        while ((count = origin.read(data, 0, 1024)) != -1) {
             out.write(data, 0, count);
         }
         IOUtils.closeQuietly(origin);
@@ -130,16 +129,16 @@ public class TestCompression
     }
 
     @Test
-    public void snappyDecompress() throws IOException
-    {
+    public void snappyDecompress() throws IOException {
         // decompress normally.
-        SnappyInputStream is = new SnappyInputStream(new BufferedInputStream(new FileInputStream("/tmp/test0.snp")));
+        SnappyInputStream is =
+                new SnappyInputStream(
+                        new BufferedInputStream(new FileInputStream("/tmp/test0.snp")));
         byte d[] = new byte[1024];
         FileOutputStream fos = new FileOutputStream("/tmp/compress-test-out-1.txt");
         BufferedOutputStream dest1 = new BufferedOutputStream(fos, 1024);
         int c;
-        while ((c = is.read(d, 0, 1024)) != -1)
-        {
+        while ((c = is.read(d, 0, 1024)) != -1) {
             dest1.write(d, 0, c);
         }
         IOUtils.closeQuietly(dest1);
@@ -151,15 +150,16 @@ public class TestCompression
     }
 
     @Test
-    public void compress() throws IOException
-    {
+    public void compress() throws IOException {
         SnappyCompression compress = new SnappyCompression();
         File file = new File(new File("/tmp/compress-test.txt"), "r");
-        long chunkSize = 5L*1024*1024;
-        Iterator<byte[]> it = compress.compress(new AbstractBackupPath.RafInputStream(new RandomAccessFile(file, "r")), chunkSize);
+        long chunkSize = 5L * 1024 * 1024;
+        Iterator<byte[]> it =
+                compress.compress(
+                        new AbstractBackupPath.RafInputStream(new RandomAccessFile(file, "r")),
+                        chunkSize);
         FileOutputStream ostream = new FileOutputStream("/tmp/test1.snp");
-        while (it.hasNext())
-        {
+        while (it.hasNext()) {
             byte[] chunk = it.next();
             ostream.write(chunk);
         }
@@ -168,10 +168,11 @@ public class TestCompression
     }
 
     @Test
-    public void decompress() throws IOException
-    {
+    public void decompress() throws IOException {
         SnappyCompression compress = new SnappyCompression();
-        compress.decompressAndClose(new FileInputStream("/tmp/test1.snp"), new FileOutputStream("/tmp/compress-test-out-2.txt"));
+        compress.decompressAndClose(
+                new FileInputStream("/tmp/test1.snp"),
+                new FileOutputStream("/tmp/compress-test-out-2.txt"));
         String md1 = SystemUtils.md5(new File("/tmp/compress-test.txt"));
         String md2 = SystemUtils.md5(new File("/tmp/compress-test-out-2.txt"));
         assertEquals(md1, md2);
